+++ internal/cmd/export.go
@@ -66,6 +66,9 @@ func runExport(_ *cobra.Command, _ []string) error {
 	}
 
 	client := slack.NewHTTPClient(prof.Token)
+	if state.cacheDir != "" {
+		client.SetCacheDir(state.cacheDir)
+	}
 	users := slack.NewUserCache(client)
 	channels := slack.NewChannelCache(client)